refactor middlewares registration
